# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Church Records Backend REST API
host: localhost:10120
...
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /nation:
    # binds a127 app logic to a route
    x-swagger-router-controller: nations.controller
    get:
      tags: [Nation]
      description: Returns list of nations from Nation Collection
      # used as the method name of the controller
      operationId: getAllNation
      responses:
        "200":
          description: Nation Found
          schema:
            type: array
            items:
              $ref: "#/definitions/country"
        "404":
          description: Empty Collection
          schema:
            $ref: "#/definitions/message"
        # responses may fall through to errors
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
    post:
      tags: [Nation]
      description: Creates new nation entry
      consumes:
        - application/json
      produces:
        - application/json
      operationId: createNation
      parameters:
        - name: nationdata
          required: true
          in: body
          description: a new nation details
          schema:
            $ref: "#/definitions/country"
      responses:
        "200":
          description: New Odd Created
          schema:
            $ref: "#/definitions/country"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
    delete:
      tags: [Nation]
      description: Remove all nations in database
      operationId: removeAllNation
      responses:
        "200":
          description: Successfully removed all nations
          schema:
            $ref: "#/definitions/message"
        default:
          description: Error
          schema:
            $ref: "#/definitions/message"
  /nation/{nationid}:
    x-swagger-router-controller: nations.controller
    get:
      tags: [Nation]
      description: Return single nation by given id
      operationId: getSingleNation
      parameters:
        - name: nationid
          required: true
          in: path
          type: string
      responses:
        "200":
          description: Nation Found
          schema:
            $ref: "#/definitions/country"
        "404":
          description: Nation not Found
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
    delete:
      tags: [Nation]
      description: Remove single nation by given id
      operationId: removeSingleNation
      parameters:
        - name: nationid
          required: true
          in: path
          type: string
      responses:
        "200":
          description: Successfully removed
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
    patch:
      tags: [Nation]
      description: Update by nation id
      operationId: updateSingleNation
      parameters:
        - name: nationid
          in: path
          required: true
          type: string
        - name: nation
          in: body
          description: The Updated nation info
          required: true
          schema:
            $ref: "#/definitions/country"
      responses:
        "200":
          description: Nation Found and Updated
          schema:
            $ref: "#/definitions/country"
        "404":
          description: Nation not Found
          schema:
            $ref: "#/definitions/message"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/message"
# The End of Nations
# The Start of Agegroup Marchavel
...
definitions:
...
  message:
    required:
      - success
      - message
    properties:
      success:
        type: boolean
      message:
        type: string
# Info Model Definition
...
  country:
    required:
      - nationname
      - nationvalue
    properties:
      _id:
        type: string
      nationname:
        type: string
      nationvalue:
        type: string
      created_at:
        type: string
        format: datetime
      updated_at:
        type: string
        format: datetime
# Agegroup Model Definition
...